Changed GameWorldPanel definitions to MVC pattern.
Still needs some work in a few places but will be easier once draw calls and UiTextureID are in.
