+++ OpenTESArena/src/Interface/GameWorldUiModel.cpp
@@ -3,6 +3,7 @@
 #include "../Game/ArenaClockUtils.h"
 #include "../Game/ArenaDateUtils.h"
 #include "../Game/Game.h"
+#include "../World/MapType.h"
 
 #include "components/utilities/String.h"
 
@@ -117,4 +118,131 @@ std::string GameWorldUiModel::getStatusButtonText(Game &game)
 	}();
 
 	return baseText + effectText;
+}
+
+std::string GameWorldUiModel::getPlayerPositionText(Game &game)
+{
+	const auto &exeData = game.getBinaryAssetLibrary().getExeData();
+	GameState &gameState = game.getGameState();
+	const MapDefinition &mapDef = gameState.getActiveMapDef();
+	const Player &player = gameState.getPlayer();
+
+	const MapType mapType = mapDef.getMapType();
+	const OriginalInt2 displayedCoords = [&player, mapType]()
+	{
+		const NewDouble3 absolutePlayerPosition = VoxelUtils::coordToNewPoint(player.getPosition());
+		const NewInt3 absolutePlayerVoxel = VoxelUtils::pointToVoxel(absolutePlayerPosition);
+		const NewInt2 playerVoxelXZ(absolutePlayerVoxel.x, absolutePlayerVoxel.z);
+		const OriginalInt2 originalVoxel = VoxelUtils::newVoxelToOriginalVoxel(playerVoxelXZ);
+
+		// The displayed coordinates in the wilderness behave differently in the original
+		// game due to how the 128x128 grid shifts to keep the player roughly centered.
+		if (mapType != MapType::Wilderness)
+		{
+			return originalVoxel;
+		}
+		else
+		{
+			const int halfWidth = RMDFile::WIDTH / 2;
+			const int halfDepth = RMDFile::DEPTH / 2;
+			return OriginalInt2(
+				halfWidth + ((originalVoxel.x + halfWidth) % RMDFile::WIDTH),
+				halfDepth + ((originalVoxel.y + halfDepth) % RMDFile::DEPTH));
+		}
+	}();
+
+	std::string str = exeData.ui.currentWorldPosition;
+
+	// Replace first %d with X, second %d with Y.
+	size_t index = str.find("%d");
+	str.replace(index, 2, std::to_string(displayedCoords.x));
+
+	index = str.find("%d", index);
+	str.replace(index, 2, std::to_string(displayedCoords.y));
+
+	return str;
+}
+
+std::string GameWorldUiModel::getCharacterSheetTooltipText()
+{
+	return "Character Sheet";
+}
+
+std::string GameWorldUiModel::getWeaponTooltipText()
+{
+	return "Draw/Sheathe Weapon";
+}
+
+std::string GameWorldUiModel::getMapTooltipText()
+{
+	return "Automap/World Map";
+}
+
+std::string GameWorldUiModel::getStealTooltipText()
+{
+	return "Steal";
+}
+
+std::string GameWorldUiModel::getStatusTooltipText()
+{
+	return "Status";
+}
+
+std::string GameWorldUiModel::getMagicTooltipText()
+{
+	return "Spells";
+}
+
+std::string GameWorldUiModel::getLogbookTooltipText()
+{
+	return "Logbook";
+}
+
+std::string GameWorldUiModel::getUseItemTooltipText()
+{
+	return "Use Item";
+}
+
+std::string GameWorldUiModel::getCampTooltipText()
+{
+	return "Camp";
+}
+
+void GameWorldUiModel::setFreeLookActive(Game &game, bool active)
+{
+	// Set relative mouse mode. When enabled, this freezes the hardware cursor in place but relative motion
+	// events are still recorded.
+	auto &inputManager = game.getInputManager();
+	inputManager.setRelativeMouseMode(active);
+
+	auto &renderer = game.getRenderer();
+	const Int2 windowDims = renderer.getWindowDimensions();
+	renderer.warpMouse(windowDims.x / 2, windowDims.y / 2);
+}
+
+VoxelDouble3 GameWorldUiModel::screenToWorldRayDirection(Game &game, const Int2 &windowPoint)
+{
+	const auto &options = game.getOptions();
+	const auto &renderer = game.getRenderer();
+	const Int2 windowDims = renderer.getWindowDimensions();
+	const int viewWidth = windowDims.x;
+	const int viewHeight = renderer.getViewHeight();
+	const double viewAspectRatio = static_cast<double>(viewWidth) / static_cast<double>(viewHeight);
+
+	auto &gameState = game.getGameState();
+	const auto &player = gameState.getPlayer();
+	const Double3 &cameraDirection = player.getDirection();
+
+	// Mouse position percents across the screen. Add 0.50 to sample at the center
+	// of the pixel.
+	const double mouseXPercent = (static_cast<double>(windowPoint.x) + 0.50) / static_cast<double>(viewWidth);
+	const double mouseYPercent = (static_cast<double>(windowPoint.y) + 0.50) / static_cast<double>(viewHeight);
+
+	return renderer.screenPointToRay(mouseXPercent, mouseYPercent, cameraDirection,
+		options.getGraphics_VerticalFOV(), viewAspectRatio);
+}
+
+Radians GameWorldUiModel::getCompassAngle(const VoxelDouble2 &direction)
+{
+	return std::atan2(-direction.y, -direction.x);
 }
